fix: apply LLM-generated title from CLI insights path
Move applyGeneratedTitle to cli/analysis/analysis-db.ts so both
the CLI hook path and the server API route share the same function.
Previously, only the dashboard-triggered analysis updated the session
title — the CLI/hook path silently skipped it.

Entire-Checkpoint: 059fb2f0ca72
